DATAMONGO-1835 - Update Javadoc and move types around.
